For metadata JSON feature, also save URIs for contained HTML files
There is a check which makes it so this feature only records URIs when
the HTML file is at the top level. E.g. if you zip an HTML file, then
a scan of the ZIP will no longer record URIs for the HTML within the
ZIP.

I think this is a mistake and I have removed that check.